@@ -8,7 +8,7 @@ function nameEncoder(value) {
   return value
     .toString()
     .split("")
-    .map(function(c) {
+    .map(function (c) {
       return validCookieNameChars.indexOf(c) > -1 ? c : encodeURIComponent(c)
     })
     .join("")
@@ -18,7 +18,7 @@ function valueEncoder(value) {
   return value
     .toString()
     .split("")
-    .map(function(c) {
+    .map(function (c) {
       return validCookieValueChars.indexOf(c) > -1 ? c : encodeURIComponent(c)
     })
     .join("")
@@ -26,7 +26,7 @@ function valueEncoder(value) {
 
 function writeCookie(name, value, attributes) {
   var attrs = Object.keys(attributes)
-    .map(function(k) {
+    .map(function (k) {
       return k + "=" + attributes[k]
     })
     .join(";")
@@ -35,7 +35,7 @@ function writeCookie(name, value, attributes) {
 
 function readCookieEffect(dispatch, props) {
   var cookies = document.cookie.split("; ")
-  var cookie = cookies.find(function(c) {
+  var cookie = cookies.find(function (c) {
     return c.substr(0, c.indexOf("=")) === props.nameEncoder(props.name)
   })
   if (cookie) {
@@ -96,7 +96,7 @@ export function ReadCookie(props) {
         converter:
           props.converter || props.json
             ? JSON.parse
-            : function(v) {
+            : function (v) {
                 return v
               },
         decoder: props.decoder || decodeURIComponent
@@ -142,7 +142,7 @@ export function WriteCookie(props) {
         converter:
           props.converter || props.json
             ? JSON.stringify
-            : function(v) {
+            : function (v) {
                 return v
               }
       },
