@@ -1,29 +1,37 @@
+import { jest } from "@jest/globals"
 import { runFx } from "../utils"
 import { Now, Delay } from "../../src"
 
 describe("Now effect", () => {
   it("should get the current time", () => {
+    jest.useFakeTimers()
     const timestamp = 9001
-    global.performance.now = () => timestamp
-
-    const action = jest.fn()
-    const timeFx = Now({ action })
-    const { dispatch } = runFx(timeFx)
-    expect(dispatch).toBeCalledWith(action, timestamp)
-
-    delete global.performance.now
+    const defaultPerformance = global.performance
+    global.performance = { now: () => timestamp }
+    try {
+      const action = jest.fn()
+      const timeFx = Now({ action })
+      const { dispatch } = runFx(timeFx)
+      expect(dispatch).toBeCalledWith(action, timestamp)
+    } finally {
+      global.performance = defaultPerformance
+      jest.useRealTimers()
+    }
   })
   it("should get the current date", () => {
     const timestamp = 9001
     const defaultDate = global.Date
-    global.Date = () => ({ timestamp })
-
-    const action = jest.fn()
-    const timeFx = Now({ asDate: true, action })
-    const { dispatch } = runFx(timeFx)
-    expect(dispatch).toBeCalledWith(action, { timestamp })
-
-    global.Date = defaultDate
+    global.Date = function () {
+      return { timestamp }
+    }
+    try {
+      const action = jest.fn()
+      const timeFx = Now({ asDate: true, action })
+      const { dispatch } = runFx(timeFx)
+      expect(dispatch).toBeCalledWith(action, { timestamp })
+    } finally {
+      global.Date = defaultDate
+    }
   })
 })
 
@@ -31,7 +39,8 @@ describe("Delay effect", () => {
   it("should get the current time after a delay", () => {
     jest.useFakeTimers()
     const timestamp = 666
-    global.performance.now = () => timestamp
+    const defaultPerformance = global.performance
+    global.performance = { now: () => timestamp }
     try {
       const action = jest.fn()
       const timeFx = Delay({ wait: timestamp, action })
@@ -40,7 +49,7 @@ describe("Delay effect", () => {
       jest.runAllTimers()
       expect(dispatch).toBeCalledWith(action, timestamp)
     } finally {
-      delete global.performance.now
+      global.performance = defaultPerformance
       jest.useRealTimers()
     }
   })
@@ -48,7 +57,9 @@ describe("Delay effect", () => {
     jest.useFakeTimers()
     const timestamp = 666
     const defaultDate = global.Date
-    global.Date = () => ({ timestamp })
+    global.Date = function () {
+      return { timestamp }
+    }
     try {
       const action = jest.fn()
       const timeFx = Delay({ wait: timestamp, asDate: true, action })
