@@ -1,14 +1,15 @@
+import { jest } from "@jest/globals"
 import { runFx } from "../utils"
 import { WatchPosition } from "../../src"
 
 describe("WatchPosition effect", () => {
   it("should return coords on success", () => {
     navigator.geolocation = {
-      watchPosition: function(successCallback) {
+      watchPosition: function (successCallback) {
         successCallback({ coords: { longitude: 42.0, latitude: 42.1 } })
         return 12
       },
-      clearWatch: function(watchId) {
+      clearWatch: function (watchId) {
         expect(watchId).toBe(12)
       }
     }
@@ -22,11 +23,11 @@ describe("WatchPosition effect", () => {
   })
   it("should call error handler on error", () => {
     navigator.geolocation = {
-      watchPosition: function(successCallback, errorCallback) {
+      watchPosition: function (successCallback, errorCallback) {
         errorCallback("things went wrong")
         return 34
       },
-      clearWatch: function(watchId) {
+      clearWatch: function (watchId) {
         expect(watchId).toBe(34)
       }
     }
